ow_session (overworld v2.0.1)

Types

id/0

-type id() :: pos_integer().

mfargs/0

-type mfargs() :: {atom(), atom(), list()}.

proxy_pid/0

-type proxy_pid() :: pid() | undefined.

serializer/0

-type serializer() :: undefined | protobuf.

token/0

-type token() :: binary() | undefined.

zone_pid/0

-type zone_pid() :: pid() | undefined.

Functions

active(EventType, EventContent, Session)

callback_mode()

connect(PID)

-spec connect(pid()) -> {ok, proxy_pid()}.

Alias for proxy(PID, self())

connected(EventType, EventContent, Session)

disconnect(PID)

-spec disconnect(pid()) -> {ok, proxy_pid()}.

Alias for proxy(PID, undefined)

disconnect_callback(PID)

-spec disconnect_callback(pid()) -> undefined | mfargs().

Get the termination callback

disconnect_callback(Callback, PID)

-spec disconnect_callback(mfargs(), pid()) -> {ok, mfargs()}.

Set the termination callback

disconnected(EventType, EventContent, Session)

game_data(PID)

-spec game_data(pid()) -> any().

Get the game data

game_data(Data, PID)

-spec game_data(any(), pid()) -> {ok, any()}.

Set the game data

id(PID)

-spec id(pid()) -> id() | undefined.

Get the session id of this session (network serializable)

id(PID, SessionID)

-spec id(pid(), id()) -> {ok, id()}.

Set the session ID for this session (network serializable)

init(_)

latency(PID)

-spec latency(pid()) -> non_neg_integer().

Get the session latency

latency(Latency, PID)

-spec latency(pos_integer(), pid()) -> {ok, pos_integer()}.

Set the session latency

proxy(PID)

-spec proxy(pid()) -> proxy_pid().

Get the pid of the session's proxy process.

proxy(PID, ProxyPID)

-spec proxy(pid(), proxy_pid()) -> {ok, proxy_pid()}.

Set the pid of the session's proxy process, including ENet, WebSocket and internal handlers.

serializer(PID)

-spec serializer(pid()) -> serializer() | undefined.

Get the format for serializing data.

serializer(Serializer, PID)

-spec serializer(serializer(), pid()) -> {ok, serializer()}.

Set the format for serializing data. If communication happens all within Erlang node(s), then there is no need to set a serializer.

start()

-spec start() -> {ok, pid()}.

Start the session server and create a new session with this ID

start(Config)

-spec start([tuple()]) -> {ok, pid()}.

stop(Pid)

-spec stop(pid()) -> ok.

Stop the session server

terminate(Reason, State, Data)

token(PID)

-spec token(pid()) -> token().

Get the session token.

token(Token, PID)

-spec token(token(), pid()) -> {ok, token()}.

Sets the session token

zone(PID)

-spec zone(pid()) -> zone_pid().

Get the zone pid

zone(ZonePID, PID)

-spec zone(zone_pid(), pid()) -> {ok, pid()}.

Sets the zone pid