ow_token_serv (overworld v2.0.1)

Types

token/0

-type token() :: binary().

Functions

code_change(OldVsn, State, Extra)

exchange(Token)

-spec exchange(token()) -> {pid(), token()} | false | denied.

Exchanges a token for a new one if it matches an existing Session PID.

handle_call(_, From, State)

handle_cast(Msg, State)

handle_info(Info, State)

init(_)

new(SessionPID)

-spec new(pid()) -> token().

Generates a new token for the given session PID.

start_link()

-spec start_link() -> {ok, pid()} | {error, term()}.

Starts the token server.

terminate(Reason, State)