ow_util (overworld v2.0.1)

Functions

any_in_list(L1, L2)

-spec any_in_list(list(), list()) -> list().

mapfind(K, V, L)

-spec mapfind(term(), term(), list()) -> term() | false.

maplistsort(Key, L)

-spec maplistsort(term(), list()) -> list().

maplistsort(Key, L, _)

mapstore(Key, Val, L, Map)

-spec mapstore(term(), term(), list(), map()) -> [map()].

pipeline(Input, Funs)

-spec pipeline(any(), list()) -> any().

remove_dups(T)

-spec remove_dups(list()) -> list().

run_bchecks(T)

-spec run_bchecks(list()) -> ok | any().

run_checks(T)

-spec run_checks(list()) -> ok | any().

take_random(List)

-spec take_random(list()) -> list().

This function will take a list of items and return a random element Alternatively an integer number of items may be specified

take_random(List, N)

-spec take_random(list(), pos_integer()) -> list().

weighted_random(Tail)

-spec weighted_random([{term(), integer()}]) -> [term()].