ow_websocket (overworld v2.0.1)

Types

ws_result/0

-type ws_result() :: {ok, any()} | {reply, [binary(), ...], any()}.

Functions

init(Req, St0)

-spec init(cowboy_req:req(), any()) -> {cowboy_websocket, cowboy_req:req(), any()}.

Handler is initialized for any new connection and logs the foreign IP

terminate(Reason, Req, Pid)

-spec terminate(any(), cowboy_req:req(), pid()) -> ok.

Terminate callback for cleanup processes

websocket_handle(Frame, Pid)

-spec websocket_handle({binary, [binary(), ...]}, pos_integer()) -> ws_result().

The websocket handler passes any binary message down to the protocol decoder for further processing. All other messages are discarded.

websocket_info(Info, SessionPid)

-spec websocket_info({From, client_msg, Msg}, SessionPid) -> Reply
                        when
                            From :: pid(),
                            Msg :: binary(),
                            SessionPid :: pid(),
                            Reply :: {reply, {binary, Msg1}, SessionPid1},
                            Msg1 :: binary(),
                            SessionPid1 :: pid().

websocket_info is triggered when a message from another erlang process comes into this handler process.

websocket_init(Pid)

-spec websocket_init(any()) -> {ok, any()}.

Set up the initial state of the websocket handler